core package refactored
