--- src/main/kotlin/io/github/janbarari/gradle/analytics/core/Stage.kt
@@ -1,5 +1,0 @@
-package io.github.janbarari.gradle.analytics.core
-
-interface Stage<I, O> {
-    fun process(input: I): O
-}
