--- src/main/kotlin/io/github/janbarari/gradle/analytics/core/UseCaseNoInput.kt
@@ -1,9 +1,0 @@
-package io.github.janbarari.gradle.analytics.core
-
-import io.github.janbarari.gradle.ExcludeJacocoGenerated
-
-@SuppressWarnings("UnnecessaryAbstractClass")
-@ExcludeJacocoGenerated
-abstract class UseCaseNoInput<OUTPUT> {
-    abstract fun execute(): OUTPUT
-}
