--- src/main/kotlin/io/github/janbarari/gradle/analytics/core/UseCaseNoOutput.kt
@@ -1,9 +1,0 @@
-package io.github.janbarari.gradle.analytics.core
-
-import io.github.janbarari.gradle.ExcludeJacocoGenerated
-
-@SuppressWarnings("UnnecessaryAbstractClass")
-@ExcludeJacocoGenerated
-abstract class UseCaseNoOutput<INPUT> {
-    abstract fun execute(input: INPUT)
-}
