+++ src/main/kotlin/io/github/janbarari/gradle/analytics/domain/usecase/SaveTemporaryMetricUseCase.kt
@@ -1,6 +1,6 @@
 package io.github.janbarari.gradle.analytics.domain.usecase
 
-import io.github.janbarari.gradle.analytics.core.UseCase
+import io.github.janbarari.gradle.core.UseCase
 import io.github.janbarari.gradle.analytics.domain.model.BuildMetric
 import io.github.janbarari.gradle.analytics.domain.repository.DatabaseRepository
