+++ src/main/kotlin/io/github/janbarari/gradle/analytics/metric/initialization/InitializationMetricRenderStage.kt
@@ -1,6 +1,6 @@
 package io.github.janbarari.gradle.analytics.metric.initialization
 
-import io.github.janbarari.gradle.analytics.core.Stage
+import io.github.janbarari.gradle.core.Stage
 import io.github.janbarari.gradle.analytics.domain.model.AnalyticsReport
 import io.github.janbarari.gradle.extension.isNull
 import io.github.janbarari.gradle.utils.MathUtils
