+++ src/main/kotlin/io/github/janbarari/gradle/analytics/metric/initialization/InitializationMetricUseCase.kt
@@ -1,6 +1,6 @@
 package io.github.janbarari.gradle.analytics.metric.initialization
 
-import io.github.janbarari.gradle.analytics.core.UseCase
+import io.github.janbarari.gradle.core.UseCase
 import io.github.janbarari.gradle.analytics.domain.model.InitializationMetric
 
 class InitializationMetricUseCase: UseCase<Long, InitializationMetric>() {
