+++ src/main/kotlin/io/github/janbarari/gradle/analytics/reporttask/render/ReportRenderPipeline.kt
@@ -1,6 +1,6 @@
 package io.github.janbarari.gradle.analytics.reporttask.render
 
-import io.github.janbarari.gradle.analytics.core.Pipeline
-import io.github.janbarari.gradle.analytics.core.Stage
+import io.github.janbarari.gradle.core.Pipeline
+import io.github.janbarari.gradle.core.Stage
 
 class ReportRenderPipeline(firstStage: Stage<String, String>): Pipeline<String, String>(firstStage)
